# Clippy: the performance parts