Hello! I've applied the fixes for anything that could be performance related and/or could make our code better.

:rotating_light: Also there was this clippy warning that sounds important but I don't know much about our Hashing strategies, please give it a look :wink: https://rust-lang.github.io/rust-clippy/master/index.html#implicit_hasher :rotating_light: 

## Performance

**parse**                   
```
time:   [5.1495 us 5.1672 us 5.1921 us]
```
**serde**
```
time:   [864.92 ns 898.90 ns 940.50 ns]
```
**edn**:
```
time:   [3.4170 us 3.4894 us 3.5709 us]
```