@naomijub 2020-09-15
It is nice to put `before` and `after` in benchmarks like this